[refactor]: 패키징, 네이밍 수정
