+++ app/src/main/java/com/pickpoint/pickpoint/ui/common/util/Utils.kt
@@ -10,7 +10,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.runtime.snapshots.SnapshotStateMap
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import com.pickpoint.pickpoint.ui.theme.PointColors
@@ -24,24 +23,3 @@ fun PointColors.getPointColorList(): List<Color>{
     )
 }
 
-// 화면 터치가 발생할 때 타이머 시작을 담당하는 composable
-@Composable
-fun timerStartHandler(
-    pointsToStart: Int = 2, // 시작하기 위한 포인트 개수
-    timeToStart: Long = 2000, // 시작하기 위한 시간 (ms)
-): Pair<SnapshotStateMap<Long, Pair<Offset, Color>>, SnapshotStateList<Pair<Offset, Color>>> {
-    val touchPoints = remember { mutableStateMapOf<Long, Pair<Offset, Color>>() }
-    val finalPoints = remember { mutableStateListOf<Pair<Offset, Color>>() }
-    var isCountingDown by remember { mutableStateOf(false) }
-
-    LaunchedEffect (touchPoints.keys.toSet()) {
-        if (touchPoints.size >= pointsToStart && !isCountingDown) {
-            isCountingDown = true
-            delay(timeToStart)
-            finalPoints.clear()
-            finalPoints.addAll(touchPoints.values)
-        }
-    }
-
-    return touchPoints to finalPoints
-}
